requirements.txt
Django>=2.0
psycopg2
coverage
coveralls
djongo
djangorestframework==3.8.2
model_mommy
